Update file structure and make function calls correct
